Rename util file to input_provider
