+++ demos/gan-lab/gan_lab_input_providers.ts
